+++ citadel-internal-service-connector/Cargo.toml
@@ -7,6 +7,7 @@ edition = "2021"
 
 [dependencies]
 citadel-internal-service-types = { workspace = true }
+citadel_logging = { workspace = true }
 tokio = { workspace = true, features = ["net", "rt", "macros"] }
 tokio-util = { workspace = true, features = ["codec"] }
 bincode2 = { workspace = true }
