+++ citadel-internal-service-connector/src/lib.rs
@@ -1,2 +1,3 @@
 pub mod codec;
+pub mod connector;
 pub mod util;
